평점 데이터타입 수정(int>double)

+++ movieJava/src/movie/model/vo/MovieVO.java
@@ -17,7 +17,7 @@ public class MovieVO {
 	}
 
 	public MovieVO(String m_code, String m_title, String m_genre, String m_director, String m_date, String m_country,
-			String m_image, String m_summary, String m_rating, int m_grade, String status) {
+			String m_image, String m_summary, String m_rating, double m_grade, String status) {
 		super();
 		this.m_code = m_code;
 		this.m_title = m_title;
@@ -108,7 +108,7 @@ public class MovieVO {
 		return m_grade;
 	}
 
-	public void setM_grade(int m_grade) {
+	public void setM_grade(double m_grade) {
 		this.m_grade = m_grade;
 	}
